requirements.txt
slack-bolt==1.16.1
slack-sdk==3.19.5
python-dotenv==0.21.0
openai==0.27.8
requests==2.28.2
pytest==7.3.1
pytest-mock==3.10.0
aiohttp==3.8.4
schedule==1.1.0
pydantic==1.10.8 